Set argv[0] to wrapped executable path, to fix wrapping Python
Set argv[0] to the full path to the wrapped executable in order to fix
the value of sys.executable when wrapping Python. This is needed since
various programs rely on value of os.path.realpath(sys.executable), and
having python-exec there is not helpful. Furthermore, Linux forces this
for scripts anyway, so it guarantees more consistent behavior across
platforms. Also add a regression test for it.

Bug: https://bugs.gentoo.org/show_bug.cgi?id=574970