Remove the unused & semi-broken 'env' tool fallback from exec()
Remove the fallback to 'env' tool on ENOEXEC from exec() since it is
most likely unnecessary (never reproduced a case for it) and shifts argv
below index 0 if a script is linked directly to the C wrapper. If it
becomes necessary in the future, it needs to be reimplemented better.