Report non-fatal errors only if stderr is a tty
Avoid reporting non-fatal errors to stderr when it's clearly redirected.
Since python-exec is used to wrap various scripts, we do not want to
alter their output if it may be in the process of being captured.

Reporting fatal errors is acceptable since the executable is not going
to be run anyway, and so there is no risk in mixing the program output
with ours.